refactor: createIdx를 handleState 영역으로 이동, addTodo()함수를 addStateToStore()로 함수 이름 바꾸고 내부 중복 로직을 setStateAndDispatch 함수로 추출
