+++ js/handleState.js
@@ -6,4 +6,21 @@ const subscribe = () => {
   return JSON.parse(localStorage.getItem("store"));
 };
 
-export { dispatch, subscribe };
+const createIdx = () => {
+  if (subscribe()) {
+    const store = subscribe();
+    if (store.length === 0) return 0;
+    const idxArray = [];
+    store.map((item) => {
+      idxArray.push(item.idx);
+    });
+    const max = idxArray.reduce(function (previous, current) {
+      return previous > current ? previous : current;
+    });
+    return max + 1;
+  } else {
+    return 0;
+  }
+};
+
+export { dispatch, subscribe, createIdx };
